[BREAKING CHANGE] restructure Service-Classes * move services to their own folder * create AbstractService class to better test if they are real services and make shure to implement the right methods * hopefully fix all calls to the new services * new autoloader-dump is mandatory
