+++ src/Service/AbstractService.php
@@ -1,0 +1,32 @@
+<?php
+/** Freesewing\DraftService class */
+namespace Freesewing\Service;
+
+/**
+ * Handles the draft service, which drafts patterns.
+ *
+ * @author    Joost De Cock <joost@decock.org>
+ * @copyright 2016 Joost De Cock
+ * @license   http://opensource.org/licenses/GPL-3.0 GNU General Public License, Version 3
+ */
+abstract class AbstractService
+{
+
+    /**
+     * Returns the name of the service
+     *
+     * This is used to load the default theme for the service when no theme is specified
+     *
+     * @see Context::loadTheme()
+     *
+     * @return string
+     */
+    abstract public function getServiceName();
+
+    /**
+     * @param \Freesewing\Context $context
+     *
+     * @return void
+     */
+    abstract public function run(\Freesewing\Context $context);
+}
